% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_load.R
\name{data_load}
\alias{data_load}
\title{Create the Data frame}
\usage{
data_load(url, sperator = ";")
}
\arguments{
\item{url}{The URL (in string) from where to download the data}

\item{sperator}{The separator (in string) used in the file}
}
\value{
A data frame with separated columns
}
\description{
Read the data set from the URL and return it as a data frame
}
\examples{
data_load(
"https://raw.githubusercontent.com/kashish1928/white_wine_dataset/main/winequality-white.csv",
";")
}
