% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist_plot.R
\name{hist_plot}
\alias{hist_plot}
\title{Plot a histogram for the count of each categorical variable}
\usage{
hist_plot(data, grp_var)
}
\arguments{
\item{data}{The dataframe includes the grp_var variable}

\item{grp_var}{The categorical column which is used for grouping other variables}
}
\value{
A histogram that shows the count of each categorical variable
}
\description{
Plot a histogram for the count of each categorical variable
}
\examples{
hist_plot(iris, "Species")
}
