% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_build.R
\name{model_build}
\alias{model_build}
\title{KNN Model Building}
\usage{
model_build(train_data, recipe, col)
}
\arguments{
\item{train_data}{The training data for the model}

\item{recipe}{The recipe for the data to build model}

\item{pred}{The column name of the response variable}
}
\value{
A fitted KNN model using the most accurate K value
}
\description{
Build the KNN model with the highest accuracy K value
}
\examples{
model_build(iris, recipes:recipe(Species ~., data = iris))
}
