% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_results.R
\name{model_results}
\alias{model_results}
\title{KNN model confusion matrix}
\usage{
model_results(test_data, model, col)
}
\arguments{
\item{test_data}{The testing data for the model}

\item{model}{The fitted KNN model}

\item{col}{The column name of the response variable}
}
\value{
A confusion matrix of the model
}
\description{
Create the confusion matrix of the fitted KNN model to assess performance
}
\examples{
split <- rsample::initial_split(iris, prop = 0.75, strata = Species)
ex_test <- rsample::testing(split)
ex_train <- rsample::training(split)
ex_recipe <- recipes::recipe(Species ~., data = ex_train)
ex_model <- model_build(ex_train, ex_recipe, "Species")
model_results(ex_test, ex_model, "Species")
}
