% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wineclassify-package.R
\docType{package}
\name{wineclassify-package}
\alias{wineclassify}
\alias{wineclassify-package}
\title{wineclassify: Create the descriptive analysis and classification model of the wine dataset}
\description{
This package includes the functions that can conduct the descriptive analysis by creating the histogram of the count of categorical variables and the correlation matrix. It also provides users to create a KNN classification model and assess the model's performance.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://dsci-310.github.io/dsci-310-group-02-pkg/}
}

}
\author{
\strong{Maintainer}: Eric HUANG \email{erichsy@student.ubc.ca}

Authors:
\itemize{
  \item Kashish Joshipura \email{kashishjoshipura@gmail.com}
  \item Peter Lee \email{leepeter718@gmail.com}
}

}
\keyword{internal}
